% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_kmer_if_not_exists}
\alias{add_kmer_if_not_exists}
\title{Add k-mer to a \code{hash map} if it does not exist}
\arguments{
\item{kmers}{\code{hash map} containing k-mers (see \link{update_kmers})}

\item{kmer}{encoded \code{integer} vector representing a k-mer}

\item{num2str}{\code{hash map} representing encoding between \code{integer} and \code{string}}

\item{kmer_decorator}{\code{function} that can add extra characters to the \code{string} representation of k-mer}
}
\description{
Add k-mer to a \code{hash map} if it does not exist
}
