% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adj_matrix}
\alias{adj_matrix}
\title{Adjacency Matrix}
\usage{
adj_matrix(x, y, states = NULL)
}
\arguments{
\item{x}{\code{NumericVector}.}

\item{y}{\code{NumericVector}.}

\item{states}{Optional integer. Specify the max number of states
between the two vectors. The outputed matrix will be of
size \code{states ^ 2 x states ^ 2}. Default: \code{max(a, b)}.}
}
\value{
A \code{S3} matrix of class \code{adj_mat}. Can be plotted with the
normal \code{plot} function.
}
\description{
Create an adjacency matrix from two ordinal variables.
}
\details{
The adjacency matrix represents paired transitions for two
variables. If we consider each possible unique pair as
serialized on an ordinal scale (\code{1, 2, .., K}) then the
column numbers of the adjacency matrix denote the start
position of the transition (the observed pair at
observation \code{i}) and the row numbers denote the end
position of the transition (the observed pair at
observation \code{i + 1}). This can be visualized graphically
with \code{plot} or as an \code{igraph} adjacency graph with
\link[igraph:graph.adjacency]{igraph::graph.adjacency}.
}
\section{Warning}{
 \code{adj_matrix} assumes that the target
variables are indexed at 0.
}

\examples{
x <- c(1, 2, 2)
y <- c(2, 1, 2)
adj_matrix(x, y) # `states` will be set to 3

}
