% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_gen.R
\name{amino_acids}
\alias{amino_acids}
\title{Amino acids}
\usage{
amino_acids()
}
\value{
A dataframe which contains basic information about all of amino acids. 
Every amino acid is described by it's name, short name and letter code.
}
\description{
Amino acids
}
