% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bubbleplot}
\alias{bubbleplot}
\title{Directed Adjacency Graphs}
\usage{
bubbleplot(x, y, data = NULL, ...)
}
\arguments{
\item{x}{\code{NumericVector}.}

\item{y}{\code{NumericVector}.}

\item{data}{Optional \code{data.frame} containing the variables to be graphed.}

\item{...}{Additional parameters passed to \code{plot}.}
}
\value{
Outputs a plot!
}
\description{
Creates "bubble plots" (\emph{i.e.}, directed adjacency graphs) showing
transitions for observed pairs of two variables.
}
\examples{
x <- c(0, 1, 1, 0, 0, 1, 0, 1)
y <- c(1, 1, 1, 0, 0, 1, 0, 1)
bubbleplot(x, y)

}
