% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{count_k_mers}
\alias{count_k_mers}
\title{Count kmers}
\usage{
count_k_mers(string, k = 2)
}
\arguments{
\item{string}{\code{character} with the sequence}

\item{k}{\code{integer} defines length of the subsequence to be counted}
}
\value{
table with number of occurrence of all k-mers in the sequence
}
\description{
Count kmers
}
\examples{
count_k_mers("AAPGAGAYY", 2)
}
