% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_kmer_num}
\alias{count_kmer_num}
\title{Count k-mers for numeric sequences (the size of k-mer should be larger than one)}
\usage{
count_kmer_num(s, d, alphabet, pos)
}
\arguments{
\item{s}{a \code{numeric} vector representing an input sequence}

\item{d}{an \code{integer} vector representing gaps between consecutive elements of k-mer}

\item{alphabet}{a \code{numeric} vector representing valid elements of k-mer}

\item{pos}{a \code{logical} value that denotes whether positional k-mers should be generated}
}
\value{
a named vector with counts of k-mers
}
\description{
Count k-mers for numeric sequences (the size of k-mer should be larger than one)
}
\details{
K-mers that contain elements from \code{alphabet} but do not exist in the input sequence are also generated.
}
\examples{
count_kmers_str(c(1,2,3,5,3,7),
d=c(0,0),
c(1, 2, 3, 4),
pos=FALSE)
}
