% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_kmers_hashed}
\alias{count_kmers_hashed}
\title{Count k-mers hashed}
\usage{
count_kmers_hashed(encoded_sequence, k, positional_kmer, P, P_K_1, M)
}
\arguments{
\item{encoded_sequence}{an encoded sequence}

\item{k}{the length of a k-mer}

\item{positional_kmer}{a \code{boolean} determining whether a k-mer is positional}

\item{P}{a hashing base}

\item{P_K_1}{a computed value for P to the power of k-1}

\item{M}{a modulo value for a hashing function}
}
\value{
a \code{DataFrame} with k-mer counts, it has 2 columns: \code{position} (the first position in the sequence) and \code{cnt}
}
\description{
Count k-mers hashed
}
