% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_kmers_helper}
\alias{count_kmers_helper}
\title{Count k-mers}
\arguments{
\item{s}{\code{integer} vector representing encoded input sequence}

\item{d}{\code{integer} vector representing gaps in k-mer}

\item{alphabet}{\code{integer} vector representing encoded alphabet}

\item{num2str}{\code{hash map} representing the sequence elements encoding between \code{integer} and \code{string}}

\item{kmer_decorator}{\code{function} that can add extra characters in order to enhance the presentation of \code{string} k-mer}

\item{pos}{\code{logical} value representing whether to count positional k-mers}
}
\value{
\code{hash map} whose key is a \code{string} presentation of k-mer and value is the number of its occurrences
}
\description{
Counts the occurrences of k-mers (the size of k-mer should be larger than one)
}
