% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_kmers_larger_than_one}
\alias{count_kmers_larger_than_one}
\title{Count k-mers that containes more than one item}
\usage{
count_kmers_larger_than_one(m, d, alphabet, pos)
}
\arguments{
\item{m}{\code{character} matrix - each row represents one sequence}

\item{d}{an \code{integer} vector representing gaps between consecutive elements of k-mer}

\item{alphabet}{a \code{numeric} vector representing valid elements of k-mer}

\item{pos}{a \code{logical} value that denotes whether positional k-mers should be generated}
}
\value{
a named vector with counts of k-mers
}
\description{
Count k-mers that containes more than one item
}
