% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{count_mers}
\alias{count_mers}
\title{Count all mers}
\usage{
count_mers(string)
}
\arguments{
\item{string}{\code{character} with the sequence}
}
\value{
table with number of occurrence of all mers in the sequence
}
\description{
Count all mers
}
\examples{
count_mers("AAPGAGAYY")
}
