% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_multimers.R
\name{count_multimers}
\alias{count_multimers}
\title{Count k-mers of various types for a given collection of sequences}
\usage{
count_multimers(
  sequences,
  k_vector,
  kmer_alphabet = getOption("seqR_alphabet_default"),
  positional_vector = rep(getOption("seqR_positional_default"), length(k_vector)),
  kmer_gaps_list = rep(list(c()), length(k_vector)),
  with_kmer_counts = getOption("seqR_with_kmer_counts_default"),
  with_kmer_names = getOption("seqR_with_kmer_names_default"),
  batch_size = getOption("seqR_batch_size_default"),
  hash_dim = getOption("seqR_hash_dim_default"),
  verbose = getOption("seqR_verbose_default")
)
}
\description{

}
