% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_unigrams}
\alias{count_unigrams}
\title{Count unigrams}
\usage{
count_unigrams(m, alphabet, pos)
}
\arguments{
\item{m}{\code{string} matrix that contains one sequence in each row}

\item{alphabet}{\code{string} vector that contains valid elements to construct unigrams}

\item{pos}{\code{logical} vector denoting whether to count positional k-mers}
}
\value{
named \code{integer} vector with unigrams' counts
}
\description{
Count unigrams
}
