% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_kmer}
\alias{create_kmer}
\title{Create k-mer}
\arguments{
\item{s}{\code{integer} vector representing an encoded sequence}

\item{d}{\code{integer} vector representing the gaps in k-mer}

\item{begin_index}{\code{integer} representing the start of k-mer (in \code{s})}

\item{num2str}{\code{hash map} representing encoding of sequence items between \code{integer} and \code{string}}

\item{kmer_decorator}{a \code{function} that can add extra characters to k-mer (for example position information)}

\item{kmer}{\code{integer} vector representing an encoded sequence}

\item{num2str}{\code{hash map} representing encoding of sequence items between \code{integer} and \code{string}}

\item{kmer_decorator}{a \code{function} that can add extra characters to k-mer (for example position information)}
}
\value{
a \code{string} representing a result k-mer (that is used for presentation)

a \code{string} representing a result k-mer (that is used for presentation)
}
\description{
Creates k-mer (of type \code{string}) from encoded (\code{integer}) vector
based on encoding described in \code{num2str} and kmer_decorator

Creates k-mer (of type \code{string}) from encoded (\code{integer}) vector
based on encoding described in \code{num2str} and kmer_decorator
}
