% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsgA.R
\docType{data}
\name{CsgA}
\alias{CsgA}
\title{CsgA data set}
\format{
a list of 24 proteins.
}
\usage{
CsgA
}
\description{
24 reviewed sequences of CsgA proteins obtained from UniProt on 13-09-2021.
}
\keyword{datasets}
