% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decode_kmer}
\alias{decode_kmer}
\title{Decode a k-mer}
\usage{
decode_kmer(encoded_sequence, d, begin_position, decoder, positional_kmer)
}
\arguments{
\item{encoded_sequence}{an encoded \code{integer} sequence}

\item{d}{an \code{integer} vector representing gaps in a k-mer}

\item{begin_position}{the starting position in the \code{encoded_sequence}}

\item{decoder}{an \code{integer} code to \code{string} sequence item decoder}

\item{positional_kmer}{a flag indicating whether a k-mer is positional}
}
\value{
a \code{StringVector} representing a decoded k-mer
}
\description{
Decode a k-mer
}
