% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{enumerate_integer_sequence}
\alias{enumerate_integer_sequence}
\title{Enumerate integer sequence}
\usage{
enumerate_integer_sequence(sequence, alphabet)
}
\arguments{
\item{sequence}{an \code{integer} sequence}

\item{alphabet}{an alphabet containing \code{integer} items}
}
\value{
an \code{integer} vector representing enumerated sequence items
}
\description{
Enumerate integer sequence
}
