% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{enumerate_numeric_sequence}
\alias{enumerate_numeric_sequence}
\title{Enumerate numeric sequence}
\usage{
enumerate_numeric_sequence(sequence, alphabet)
}
\arguments{
\item{sequence}{an \code{numeric} sequence}

\item{alphabet}{an alphabet containing \code{numeric} items}
}
\value{
an \code{integer} vector representing enumerated sequence items
}
\description{
Enumerate numeric sequence
}
