% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{enumerate_string_sequence}
\alias{enumerate_string_sequence}
\title{Enumerate string sequence}
\usage{
enumerate_string_sequence(sequence, alphabet)
}
\arguments{
\item{sequence}{a \code{string} sequence}

\item{alphabet}{an alphabet containing \code{string} items}
}
\value{
an \code{integer} vector representing enumerated sequence items
}
\description{
Enumerate string sequence
}
