% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fill_encoded_int_vector}
\alias{fill_encoded_int_vector}
\title{Encode sequence vector (replace items to numbers)}
\arguments{
\item{str_v}{the input (Rcpp) sequence}

\item{res}{the result (encoded) vector}

\item{val2int}{the encoder}
}
\description{
Encode sequence vector (replace items to numbers)
}
\details{
\code{SEQ_TYPE} - the type of a sequence of the input sequence (Rcpp)
\code{ELEM_TYPE} - the type of an item of the input sequence (c++)
}
