% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fill_items_coding_maps}
\alias{fill_items_coding_maps}
\title{Prepare encoding and decoding \code{hash maps} for sequence items}
\arguments{
\item{elems}{the input elements of a sequence}

\item{val2num}{the reference to \code{hash map} representing the encoding to \code{integer} value}

\item{num2str}{the reference to \code{hash map} representing the (reversed to \code{val2num}) encoding to \code{string} value}

\item{lowest_not_used_num}{the reference to \code{integer} value denoting current counter used to encode elements}

\item{val2str_converter}{\code{function} that take a sequence item and returns its string representation that is used for presentation}
}
\description{
Prepare encoding and decoding \code{hash maps} for sequence items
}
\details{
Enumerates each sequence item in order to convert \code{non-integer} values to \code{integer} ones
\code{B} is the template Rcpp input type
\code{S} is the template c++ input type of one element
}
