% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_all_kmers}
\alias{generate_all_kmers}
\title{Generate all k-mers}
\usage{
generate_all_kmers(k, P, M, decoder)
}
\arguments{
\item{k}{the size of the k-mer}

\item{P}{a base for a hashing function}

\item{M}{a modulo value for a hashing function}

\item{decoder}{an \code{StringVector} representing integer-string sequence item decoding}
}
\value{
a \code{string} vector representing all k-mers
}
\description{
Generate all k-mers
}
