% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{generate_random_seq}
\alias{generate_random_seq}
\title{Generate random sequence}
\usage{
generate_random_seq(nsize = 50)
}
\arguments{
\item{nsize}{\code{integer} number of elements in the sequence}
}
\value{
Random \code{character} object of given length created from valid amino acid letters.
}
\description{
Generate random sequence
}
\examples{
generate_random_seq(10)
}
