% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_hash}
\alias{get_hash}
\title{Get hash of k-mer that is in a given sequence}
\usage{
get_hash(s, d, begin_index, pos)
}
\arguments{
\item{s}{\code{integer} vector representing input sequence}

\item{d}{\code{integer} vector representing the gaps between consecutive elements of k-mer}

\item{begin_index}{\code{integer} value representing the begin index of the k-mer}

\item{pos}{\code{logical} value representing whether k-mer is positional}
}
\value{
\code{integer} value representing the result of a hashing function
}
\description{
Get hash of k-mer that is in a given sequence
}
