% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_hash_for_word}
\alias{get_hash_for_word}
\title{Get hash of a sequence}
\usage{
get_hash_for_word(kmer)
}
\arguments{
\item{kmer}{\code{integer} vector representing a word to be hashed}
}
\value{
\code{integer} value representing the result of a hashing function
}
\description{
Get hash of a sequence
}
