% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_not_allowed_sequence_positions}
\alias{get_not_allowed_sequence_positions}
\title{Get not allowed sequence positions}
\usage{
get_not_allowed_sequence_positions(encoded_sequence)
}
\arguments{
\item{encoded_sequence}{an \code{integer} vector representing an encoded sequence}
}
\value{
an \code{integer} vector representing positions in a sequence where there are items equal to \code{NOT_ALLOWED_CHARACTER_CODE}
}
\description{
Get not allowed sequence positions
}
