% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_total_size_of_kmer}
\alias{get_total_size_of_kmer}
\title{Get the total size of k-mer}
\arguments{
\item{s}{\code{integer} vector of encoded elements of a sequence (see Details)}

\item{d}{\code{integer} vector which denotes the gaps in k-mer}

\item{begin_index}{\code{integer} representing the begin index (in \code{s}) of the k-mer}

\item{num2str}{a \code{hash map} representing the encoding between number and \code{string} representation of each alphabet's item}

\item{kmer}{\code{integer} vector representing the encoded k-mer (\link{get_total_size_of_kmer})}

\item{num2str}{\code{hash map} representing the encoding between the integer and string}
}
\value{
\code{int} denoting the total size (number of characters) of k-mer

the number of characters in the result \code{string} that is the result of decoding each \code{integer} from \code{kmer}
}
\description{
Computes the number of characters of the result k-mer
taking into account the base alphabet.

The number of characters of the result k-mer (after decoding from \code{integer} to \code{string})
}
\details{
Each element of a sequence is previously encoded to an integer in order to make hashing computation
more convenient
}
