% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_window_length}
\alias{get_window_length}
\title{Get k-mer window length}
\usage{
get_window_length(d)
}
\arguments{
\item{d}{\code{integer} vector with distances between consequent elements}
}
\value{
\code{integer} representing the total window length
}
\description{
Compute a k-mer window length. The window length is the total size 
used by the k-mer - the number of elements and the size of gaps.
}
