% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_kmer_allowed}
\alias{is_kmer_allowed}
\title{Is k-mer allowed}
\arguments{
\item{s}{\code{integer} vector of encoded sequence characters}

\item{d}{\code{integer} vector representing gaps between elements of k-mer}

\item{begin_index}{\code{integer} representing the start index of k-mer in \code{s}}

\item{is_item_allowed}{\code{hash map} that answers the question whether the element is in the alphabet}
}
\value{
\code{logical} value denoting whether k-mer is valid (contains valid characters that are in the alphabet)
}
\description{
Checks whether all elements of the given k-mer are contained in the alphabet set
}
