% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage.R
\name{ptable}
\alias{ptable}
\title{Percentage Table}
\usage{
ptable(data, complete.obs = T, reduce = F)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{complete.obs}{A boolean value. Calculate percentages based on
complete observations for each pair of variables. Default: \code{TRUE}.}

\item{reduce}{A boolean value. For each variable pair, reduce the data to
only changes before calculating the percentage table.}
}
\value{
A \code{data.frame} where the last three columns break down the
relative frequency of variable \code{X} being larger than variable
\code{Y}, vice versa, and when \code{X} == \code{Y}.
}
\description{
Construct a table of relative frequencies.
}
\examples{
x <- data.frame(a = c(0, 0, 1, 4), b = c(1, 2, 1, 0))
ptable(x)

}
