% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{count_kmers}
\alias{count_kmers}
\title{Count k-mers}
\usage{
count_kmers(seq, d, alphabet, pos)
}
\arguments{
\item{seq}{\code{string} matrix (each row of the matrix represents one sequence) or vector}

\item{d}{\code{integer} vector representing the length of gaps between consecutive items of k-mer. See details.}

\item{alphabet}{\code{string} vector representing the items which only should be used for generating k-mers}

\item{pos}{\code{logical} value representing whether positional k-mers should be count}
}
\value{
named \code{integer} vector with k-mers counts
}
\description{
Counts positional or non positional k-mers
}
\details{
Depending on the value of parameter \code{pos} positional or non positional k-mers can be taken
into account. Positional k-mers are related to its position (column index in the given \code{seq} matrix),
so, for example k-mer \code{abc} that starts on the first position is a different k-mer
than \code{abc} that starts on the second position. 
As for non positional k-mers, all \code{abc}, regardless of their position, are treated the same.
Besides, it is important to note that during the generation of k-mers the given \code{alphabet} vector
is taken into account which means that we do not count k-mers that contain elements not included in the alphabet
and additionally, if non positional k-mers are used, all possible ones to create using the given alphabet are returned,
even if their count is equal to 0.
The format of the non positional k-mer is the following:
each element is separated by . char.
The format of the positional k-mer is similar to the above one.
The only thing that differs is that the prefix index_ is added to its representation of elements.
}
