% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{seqR-package}
\alias{seqR-package}
\alias{seqR}
\title{seqR: Fast and Comprehensive K-Mer Counting Package}
\description{
The \code{seqR} package provides in-memory, probabilistic,
highly-optimized, and multi-threaded implementation of k-mer counting.
}
\author{
Jadwiga Słowik
}
